Move rstrip and test_rstrip to format and test_format.
